requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-cartesia
livekit-plugins-silero
aiohttp
twilio
python-dotenv